Saving a frame in a jpeg file

+++ include/thumbnailer.h
@@ -33,5 +33,5 @@
 #include <libavcodec/avcodec.h>
 #include <libavutil/opt.h>
 #include <libswscale/swscale.h>
-#include <jpeglib.h>
-int tve_open_video (const char *fname, int64_t second);
+
+int tve_open_video (const char *fname, int second);
